requirements.txt
alembic
black
fastapi
flake8
httpx
mypy
SQLAlchemy
psycopg2-binary
pydantic
pytest
python-dotenv
sqlalchemy
uvicorn
